refector(oat): re-organize types
