+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/storage.ts
@@ -1,3 +1,4 @@
+export * from './storage/types';
 export * from './storage/access-token.storage.factory';
 export * from './storage/authorization-code.storage.factory';
 export * from './storage/id-token.storage.factory';
