+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/storage/types.ts
@@ -1,0 +1,14 @@
+import { AccessTokenResponse } from '../types';
+
+export type StoredAccessTokenResponse<
+  T extends AccessTokenResponse = AccessTokenResponse,
+> = {
+  createdAt: number;
+  expiresAt: number;
+  response: T;
+};
+
+export type StoredRefreshToken = {
+  expiresAt: number;
+  token: string;
+};
