+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/types/state-action.types.ts
@@ -1,7 +1,7 @@
 import { ObservableInput } from 'rxjs';
 
+import { StateData } from './oauth2-services.types';
 import { AccessTokenResponse } from './standard.types';
-import { StateData } from './storages.types';
 import { RequiredOnly } from './utils.type';
 
 export type StateActionInfo<N extends string = string, T = unknown> = {
