+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/types/storages.types.ts
@@ -1,22 +1,5 @@
 import { Observable } from 'rxjs';
 
-import { AccessTokenResponse } from './standard.types';
-
-export type StoredAccessTokenResponse<
-  T extends AccessTokenResponse = AccessTokenResponse,
-> = {
-  createdAt: number;
-  expiresAt: number;
-  response: T;
-};
-
-export type StoredRefreshToken = {
-  expiresAt: number;
-  token: string;
-};
-
-export type StateData = object;
-
 export interface KeyValuePairStorage {
   loadItem<T = unknown>(key: string): Promise<T | null>;
   storeItem<T = unknown>(key: string, value: T): Promise<T>;
